refactor: use a hardcoded list of map ids instead of scanning itemframes
